requirements.txt
django-admin==2.0.1
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.0
drf-yasg==1.20.0
django-rest-swagger==2.2.0
django-cors-headers==3.13.0